game tag model simplification

+++ netlify/functions/tags/delete.ts
@@ -1,6 +1,14 @@
 import { CustomHandler } from "../../types";
+import { CustomError } from "../../utils/error";
 
 const deleteHandler: CustomHandler<"tags/delete"> = async (prisma, params) => {
+  const gamesUsingTag = await prisma.game.count({
+    where: { tags: { has: params.id } },
+  });
+  if (gamesUsingTag > 0) {
+    throw new CustomError("Tag is still in use by one or more games", 409);
+  }
+
   const deletedTag = await prisma.tags.delete({
     where: { id: params.id },
   });
